Fixed bug that crashed quanta if you set a breakpoint in iconborder without having a debugger enabled Fixed bug that quanta didnt care about breakpoint changes made in the icon border at the same line as the current line Added watches  to DBGp implementation Restructured watch handling a bit and moved gubed specific stuff to gubed

+++ components/debugger/debuggerbreakpointlist.cpp
@@ -40,7 +40,9 @@ DebuggerBreakpointList::~DebuggerBreakpointList()
 
 void DebuggerBreakpointList::add(DebuggerBreakpoint* bp)
 {
-  quantaApp->debugger()->UI()->showBreakpoint(*bp);
+  if(quantaApp->debugger()->UI())
+    quantaApp->debugger()->UI()->showBreakpoint(*bp);
+  
   m_breakpointList->push_front(bp);
   if(bp->type() == DebuggerBreakpoint::LineBreakpoint)
     quantaApp->debugger()->setMark(bp->filePath(), bp->line(), true, KTextEditor::MarkInterface::markType02);
@@ -60,7 +62,8 @@ void DebuggerBreakpointList::remove(DebuggerBreakpoint* bp)
 
     tmp = (*it);
     // Remove it from the bp-list
-    quantaApp->debugger()->UI()->deleteBreakpoint(*bp);
+    if(quantaApp->debugger()->UI())
+      quantaApp->debugger()->UI()->deleteBreakpoint(*bp);
 
     // Remove editor markpoint if there is one...
     if(bp->type() == DebuggerBreakpoint::LineBreakpoint)
